LayoutTests/platform/gtk/http/tests/security/xss-DENIED-xsl-external-entity-redirect-expected.txt

CONSOLE MESSAGE: line 1: Document is empty

This test includes a cross-origin external entity. It passes if the load fails and thus there is no text below this line.
